+++ lib/src/widgets/editor.dart
@@ -25,6 +25,8 @@ import 'delegate.dart';
 import 'image.dart';
 import 'raw_editor.dart';
 import 'text_selection.dart';
+import 'video_app.dart';
+import 'youtube_video_app.dart';
 
 const linkPrefixes = [
   'mailto:', // email
@@ -95,7 +97,8 @@ String _standardizeImageUrl(String url) {
   return url;
 }
 
-Widget _defaultEmbedBuilder(BuildContext context, leaf.Embed node) {
+Widget _defaultEmbedBuilder(
+    BuildContext context, leaf.Embed node, bool readOnly) {
   assert(!kIsWeb, 'Please provide EmbedBuilder for Web');
   switch (node.value.type) {
     case 'image':
@@ -105,6 +108,13 @@ Widget _defaultEmbedBuilder(BuildContext context, leaf.Embed node) {
           : isBase64(imageUrl)
               ? Image.memory(base64.decode(imageUrl))
               : Image.file(io.File(imageUrl));
+    case 'video':
+      final videoUrl = node.value.data;
+      if (videoUrl.contains('youtube.com') || videoUrl.contains('youtu.be')) {
+        return YoutubeVideoApp(
+            videoUrl: videoUrl, context: context, readOnly: readOnly);
+      }
+      return VideoApp(videoUrl: videoUrl, context: context, readOnly: readOnly);
     default:
       throw UnimplementedError(
         'Embeddable type "${node.value.type}" is not supported by default '
@@ -125,6 +135,7 @@ class QuillEditor extends StatefulWidget {
       required this.readOnly,
       required this.expands,
       this.showCursor,
+      this.paintCursorAboveText,
       this.placeholder,
       this.enableInteractiveSelection = true,
       this.scrollBottomInset = 0,
@@ -165,6 +176,7 @@ class QuillEditor extends StatefulWidget {
   final EdgeInsetsGeometry padding;
   final bool autoFocus;
   final bool? showCursor;
+  final bool? paintCursorAboveText;
   final bool readOnly;
   final String? placeholder;
   final bool enableInteractiveSelection;
@@ -287,7 +299,7 @@ class _QuillEditorState extends State<QuillEditor>
             width: 2,
             radius: cursorRadius,
             offset: cursorOffset,
-            paintAboveText: paintCursorAboveText,
+            paintAboveText: widget.paintCursorAboveText ?? paintCursorAboveText,
             opacityAnimates: cursorOpacityAnimates,
           ),
           widget.textCapitalization,
@@ -901,7 +913,21 @@ class RenderEditor extends RenderEditableContainerBox
   double? getOffsetToRevealCursor(
       double viewportHeight, double scrollOffset, double offsetInViewport) {
     final endpoints = getEndpointsForSelection(selection);
-    final endpoint = endpoints.first;
+
+    // when we drag the right handle, we should get the last point
+    TextSelectionPoint endpoint;
+    if (selection.isCollapsed) {
+      endpoint = endpoints.first;
+    } else {
+      if (selection is DragTextSelection) {
+        endpoint = (selection as DragTextSelection).first
+            ? endpoints.first
+            : endpoints.last;
+      } else {
+        endpoint = endpoints.first;
+      }
+    }
+
     final child = childAtPosition(selection.extent);
     const kMargin = 8.0;