+++ lib/src/widgets/simple_viewer.dart
@@ -20,10 +20,13 @@ import 'delegate.dart';
 import 'editor.dart';
 import 'text_block.dart';
 import 'text_line.dart';
+import 'video_app.dart';
+import 'youtube_video_app.dart';
 
 class QuillSimpleViewer extends StatefulWidget {
   const QuillSimpleViewer({
     required this.controller,
+    required this.readOnly,
     this.customStyles,
     this.truncate = false,
     this.truncateScale,
@@ -51,6 +54,7 @@ class QuillSimpleViewer extends StatefulWidget {
   final double scrollBottomInset;
   final EdgeInsetsGeometry padding;
   final EmbedBuilder? embedBuilder;
+  final bool readOnly;
 
   @override
   _QuillSimpleViewerState createState() => _QuillSimpleViewerState();
@@ -97,7 +101,8 @@ class _QuillSimpleViewerState extends State<QuillSimpleViewer>
 
   EmbedBuilder get embedBuilder => widget.embedBuilder ?? _defaultEmbedBuilder;
 
-  Widget _defaultEmbedBuilder(BuildContext context, leaf.Embed node) {
+  Widget _defaultEmbedBuilder(
+      BuildContext context, leaf.Embed node, bool readOnly) {
     assert(!kIsWeb, 'Please provide EmbedBuilder for Web');
     switch (node.value.type) {
       case 'image':
@@ -107,6 +112,14 @@ class _QuillSimpleViewerState extends State<QuillSimpleViewer>
             : isBase64(imageUrl)
                 ? Image.memory(base64.decode(imageUrl))
                 : Image.file(io.File(imageUrl));
+      case 'video':
+        final videoUrl = node.value.data;
+        if (videoUrl.contains('youtube.com') || videoUrl.contains('youtu.be')) {
+          return YoutubeVideoApp(
+              videoUrl: videoUrl, context: context, readOnly: readOnly);
+        }
+        return VideoApp(
+            videoUrl: videoUrl, context: context, readOnly: readOnly);
       default:
         throw UnimplementedError(
           'Embeddable type "${node.value.type}" is not supported by default '
@@ -207,7 +220,8 @@ class _QuillSimpleViewerState extends State<QuillSimpleViewer>
             embedBuilder,
             _cursorCont,
             indentLevelCounts,
-            _handleCheckboxTap);
+            _handleCheckboxTap,
+            widget.readOnly);
         result.add(editableTextBlock);
       } else {
         throw StateError('Unreachable.');
@@ -234,6 +248,7 @@ class _QuillSimpleViewerState extends State<QuillSimpleViewer>
       textDirection: _textDirection,
       embedBuilder: embedBuilder,
       styles: _styles,
+      readOnly: widget.readOnly,
     );
     final editableTextLine = EditableTextLine(
         node,