+++ lib/src/widgets/text_selection.dart
@@ -23,6 +23,41 @@ TextSelection localSelection(Node node, TextSelection selection, fromParent) {
 
 enum _TextSelectionHandlePosition { START, END }
 
+/// internal use, used to get drag direction information
+class DragTextSelection extends TextSelection {
+  const DragTextSelection({
+    TextAffinity affinity = TextAffinity.downstream,
+    int baseOffset = 0,
+    int extentOffset = 0,
+    bool isDirectional = false,
+    this.first = true,
+  }) : super(
+          baseOffset: baseOffset,
+          extentOffset: extentOffset,
+          affinity: affinity,
+          isDirectional: isDirectional,
+        );
+
+  final bool first;
+
+  @override
+  DragTextSelection copyWith({
+    int? baseOffset,
+    int? extentOffset,
+    TextAffinity? affinity,
+    bool? isDirectional,
+    bool? first,
+  }) {
+    return DragTextSelection(
+      baseOffset: baseOffset ?? this.baseOffset,
+      extentOffset: extentOffset ?? this.extentOffset,
+      affinity: affinity ?? this.affinity,
+      isDirectional: isDirectional ?? this.isDirectional,
+      first: first ?? this.first,
+    );
+  }
+}
+
 class EditorTextSelectionOverlay {
   EditorTextSelectionOverlay(
     this.value,
@@ -156,9 +191,21 @@ class EditorTextSelectionOverlay {
       default:
         throw 'Invalid position';
     }
+
+    final currSelection = newSelection != null
+        ? DragTextSelection(
+            baseOffset: newSelection.baseOffset,
+            extentOffset: newSelection.extentOffset,
+            affinity: newSelection.affinity,
+            isDirectional: newSelection.isDirectional,
+            first: position == _TextSelectionHandlePosition.START,
+          )
+        : null;
+
     selectionDelegate
-      ..textEditingValue =
-          value.copyWith(selection: newSelection, composing: TextRange.empty)
+      ..userUpdateTextEditingValue(
+          value.copyWith(selection: currSelection, composing: TextRange.empty),
+          SelectionChangedCause.drag)
       ..bringIntoView(textPosition);
   }
 
@@ -286,6 +333,8 @@ class _TextSelectionHandleOverlay extends StatefulWidget {
 class _TextSelectionHandleOverlayState
     extends State<_TextSelectionHandleOverlay>
     with SingleTickerProviderStateMixin {
+  late Offset _dragPosition;
+  late Size _handleSize;
   late AnimationController _controller;
 
   Animation<double> get _opacity => _controller.view;
@@ -324,11 +373,13 @@ class _TextSelectionHandleOverlayState
     super.dispose();
   }
 
-  void _handleDragStart(DragStartDetails details) {}
+  void _handleDragStart(DragStartDetails details) {
+    _dragPosition = details.globalPosition + Offset(0, -_handleSize.height);
+  }
 
   void _handleDragUpdate(DragUpdateDetails details) {
-    final position =
-        widget.renderObject!.getPositionForOffset(details.globalPosition);
+    _dragPosition += details.delta;
+    final position = widget.renderObject!.getPositionForOffset(_dragPosition);
     if (widget.selection.isCollapsed) {
       widget.onSelectionHandleChanged(TextSelection.fromPosition(position));
       return;
@@ -397,6 +448,7 @@ class _TextSelectionHandleOverlayState
     final handleAnchor =
         widget.selectionControls.getHandleAnchor(type!, lineHeight);
     final handleSize = widget.selectionControls.getHandleSize(lineHeight);
+    _handleSize = handleSize;
 
     final handleRect = Rect.fromLTWH(
       -handleAnchor.dx,