+++ lib/src/widgets/toolbar.dart
@@ -2,10 +2,11 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
 
 import '../models/documents/attribute.dart';
 import 'controller.dart';
+import 'toolbar/arrow_indicated_button_list.dart';
+import 'toolbar/camera_button.dart';
 import 'toolbar/clear_format_button.dart';
 import 'toolbar/color_button.dart';
 import 'toolbar/history_button.dart';
@@ -16,6 +17,7 @@ import 'toolbar/link_style_button.dart';
 import 'toolbar/select_header_style_button.dart';
 import 'toolbar/toggle_check_list_button.dart';
 import 'toolbar/toggle_style_button.dart';
+import 'toolbar/video_button.dart';
 
 export 'toolbar/clear_format_button.dart';
 export 'toolbar/color_button.dart';
@@ -30,8 +32,13 @@ export 'toolbar/select_header_style_button.dart';
 export 'toolbar/toggle_check_list_button.dart';
 export 'toolbar/toggle_style_button.dart';
 
-typedef OnImagePickCallback = Future<String> Function(File file);
-typedef ImagePickImpl = Future<String?> Function(ImageSource source);
+typedef OnImagePickCallback = Future<String?> Function(File file);
+typedef OnVideoPickCallback = Future<String?> Function(File file);
+typedef FilePickImpl = Future<String?> Function(BuildContext context);
+typedef WebImagePickImpl = Future<String?> Function(
+    OnImagePickCallback onImagePickCallback);
+typedef WebVideoPickImpl = Future<String?> Function(
+    OnVideoPickCallback onImagePickCallback);
 
 // The default size of the icon of a button.
 const double kDefaultIconSize = 18;
@@ -39,10 +46,13 @@ const double kDefaultIconSize = 18;
 // The factor of how much larger the button is in relation to the icon.
 const double kIconButtonFactor = 1.77;
 
-class QuillToolbar extends StatefulWidget implements PreferredSizeWidget {
+class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
   const QuillToolbar({
     required this.children,
     this.toolBarHeight = 36,
+    this.color,
+    this.filePickImpl,
+    this.multiRowsDisplay,
     Key? key,
   }) : super(key: key);
 
@@ -66,271 +76,258 @@ class QuillToolbar extends StatefulWidget implements PreferredSizeWidget {
     bool showLink = true,
     bool showHistory = true,
     bool showHorizontalRule = false,
+    bool multiRowsDisplay = true,
+    bool showCamera = true,
     OnImagePickCallback? onImagePickCallback,
+    OnVideoPickCallback? onVideoPickCallback,
+    FilePickImpl? filePickImpl,
+    WebImagePickImpl? webImagePickImpl,
+    WebVideoPickImpl? webVideoPickImpl,
     Key? key,
   }) {
+    final isButtonGroupShown = [
+      showHistory ||
+          showBoldButton ||
+          showItalicButton ||
+          showUnderLineButton ||
+          showStrikeThrough ||
+          showColorButton ||
+          showBackgroundColorButton ||
+          showClearFormat ||
+          onImagePickCallback != null ||
+          onVideoPickCallback != null,
+      showHeaderStyle,
+      showListNumbers || showListBullets || showListCheck || showCodeBlock,
+      showQuote || showIndent,
+      showLink || showHorizontalRule
+    ];
+
     return QuillToolbar(
-        key: key,
-        toolBarHeight: toolbarIconSize * 2,
-        children: [
-          Visibility(
-            visible: showHistory,
-            child: HistoryButton(
-              icon: Icons.undo_outlined,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              undo: true,
-            ),
+      key: key,
+      toolBarHeight: toolbarIconSize * 2,
+      multiRowsDisplay: multiRowsDisplay,
+      children: [
+        if (showHistory)
+          HistoryButton(
+            icon: Icons.undo_outlined,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            undo: true,
           ),
-          Visibility(
-            visible: showHistory,
-            child: HistoryButton(
-              icon: Icons.redo_outlined,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              undo: false,
-            ),
+        if (showHistory)
+          HistoryButton(
+            icon: Icons.redo_outlined,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            undo: false,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showBoldButton,
-            child: ToggleStyleButton(
-              attribute: Attribute.bold,
-              icon: Icons.format_bold,
-              iconSize: toolbarIconSize,
-              controller: controller,
-            ),
+        if (showBoldButton)
+          ToggleStyleButton(
+            attribute: Attribute.bold,
+            icon: Icons.format_bold,
+            iconSize: toolbarIconSize,
+            controller: controller,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showItalicButton,
-            child: ToggleStyleButton(
-              attribute: Attribute.italic,
-              icon: Icons.format_italic,
-              iconSize: toolbarIconSize,
-              controller: controller,
-            ),
+        if (showItalicButton)
+          ToggleStyleButton(
+            attribute: Attribute.italic,
+            icon: Icons.format_italic,
+            iconSize: toolbarIconSize,
+            controller: controller,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showUnderLineButton,
-            child: ToggleStyleButton(
-              attribute: Attribute.underline,
-              icon: Icons.format_underline,
-              iconSize: toolbarIconSize,
-              controller: controller,
-            ),
+        if (showUnderLineButton)
+          ToggleStyleButton(
+            attribute: Attribute.underline,
+            icon: Icons.format_underline,
+            iconSize: toolbarIconSize,
+            controller: controller,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showStrikeThrough,
-            child: ToggleStyleButton(
-              attribute: Attribute.strikeThrough,
-              icon: Icons.format_strikethrough,
-              iconSize: toolbarIconSize,
-              controller: controller,
-            ),
+        if (showStrikeThrough)
+          ToggleStyleButton(
+            attribute: Attribute.strikeThrough,
+            icon: Icons.format_strikethrough,
+            iconSize: toolbarIconSize,
+            controller: controller,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showColorButton,
-            child: ColorButton(
-              icon: Icons.color_lens,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              background: false,
-            ),
+        if (showColorButton)
+          ColorButton(
+            icon: Icons.color_lens,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            background: false,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showBackgroundColorButton,
-            child: ColorButton(
-              icon: Icons.format_color_fill,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              background: true,
-            ),
+        if (showBackgroundColorButton)
+          ColorButton(
+            icon: Icons.format_color_fill,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            background: true,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: showClearFormat,
-            child: ClearFormatButton(
-              icon: Icons.format_clear,
-              iconSize: toolbarIconSize,
-              controller: controller,
-            ),
+        if (showClearFormat)
+          ClearFormatButton(
+            icon: Icons.format_clear,
+            iconSize: toolbarIconSize,
+            controller: controller,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: onImagePickCallback != null,
-            child: ImageButton(
-              icon: Icons.image,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              imageSource: ImageSource.gallery,
-              onImagePickCallback: onImagePickCallback,
-            ),
+        if (onImagePickCallback != null)
+          ImageButton(
+            icon: Icons.image,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            onImagePickCallback: onImagePickCallback,
+            filePickImpl: filePickImpl,
+            webImagePickImpl: webImagePickImpl,
           ),
-          const SizedBox(width: 0.6),
-          Visibility(
-            visible: onImagePickCallback != null,
-            child: ImageButton(
+        if (onVideoPickCallback != null)
+          VideoButton(
+            icon: Icons.movie_creation,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            onVideoPickCallback: onVideoPickCallback,
+            filePickImpl: filePickImpl,
+            webVideoPickImpl: webImagePickImpl,
+          ),
+        if ((onImagePickCallback != null || onVideoPickCallback != null) &&
+            showCamera)
+          CameraButton(
               icon: Icons.photo_camera,
               iconSize: toolbarIconSize,
               controller: controller,
-              imageSource: ImageSource.camera,
               onImagePickCallback: onImagePickCallback,
-            ),
-          ),
-          Visibility(
-            visible: showHeaderStyle,
-            child: VerticalDivider(
-              indent: 12,
-              endIndent: 12,
-              color: Colors.grey.shade400,
-            ),
-          ),
-          Visibility(
-            visible: showHeaderStyle,
-            child: SelectHeaderStyleButton(
-              controller: controller,
-              iconSize: toolbarIconSize,
-            ),
-          ),
+              onVideoPickCallback: onVideoPickCallback,
+              filePickImpl: filePickImpl,
+              webImagePickImpl: webImagePickImpl,
+              webVideoPickImpl: webVideoPickImpl),
+        if (isButtonGroupShown[0] &&
+            (isButtonGroupShown[1] ||
+                isButtonGroupShown[2] ||
+                isButtonGroupShown[3] ||
+                isButtonGroupShown[4]))
           VerticalDivider(
             indent: 12,
             endIndent: 12,
             color: Colors.grey.shade400,
           ),
-          Visibility(
-            visible: showListNumbers,
-            child: ToggleStyleButton(
-              attribute: Attribute.ol,
-              controller: controller,
-              icon: Icons.format_list_numbered,
-              iconSize: toolbarIconSize,
-            ),
+        if (showHeaderStyle)
+          SelectHeaderStyleButton(
+            controller: controller,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: showListBullets,
-            child: ToggleStyleButton(
-              attribute: Attribute.ul,
-              controller: controller,
-              icon: Icons.format_list_bulleted,
-              iconSize: toolbarIconSize,
-            ),
+        if (isButtonGroupShown[1] &&
+            (isButtonGroupShown[2] ||
+                isButtonGroupShown[3] ||
+                isButtonGroupShown[4]))
+          VerticalDivider(
+            indent: 12,
+            endIndent: 12,
+            color: Colors.grey.shade400,
           ),
-          Visibility(
-            visible: showListCheck,
-            child: ToggleCheckListButton(
-              attribute: Attribute.unchecked,
-              controller: controller,
-              icon: Icons.check_box,
-              iconSize: toolbarIconSize,
-            ),
+        if (showListNumbers)
+          ToggleStyleButton(
+            attribute: Attribute.ol,
+            controller: controller,
+            icon: Icons.format_list_numbered,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: showCodeBlock,
-            child: ToggleStyleButton(
-              attribute: Attribute.codeBlock,
-              controller: controller,
-              icon: Icons.code,
-              iconSize: toolbarIconSize,
-            ),
+        if (showListBullets)
+          ToggleStyleButton(
+            attribute: Attribute.ul,
+            controller: controller,
+            icon: Icons.format_list_bulleted,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: !showListNumbers &&
-                !showListBullets &&
-                !showListCheck &&
-                !showCodeBlock,
-            child: VerticalDivider(
-              indent: 12,
-              endIndent: 12,
-              color: Colors.grey.shade400,
-            ),
+        if (showListCheck)
+          ToggleCheckListButton(
+            attribute: Attribute.unchecked,
+            controller: controller,
+            icon: Icons.check_box,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: showQuote,
-            child: ToggleStyleButton(
-              attribute: Attribute.blockQuote,
-              controller: controller,
-              icon: Icons.format_quote,
-              iconSize: toolbarIconSize,
-            ),
+        if (showCodeBlock)
+          ToggleStyleButton(
+            attribute: Attribute.codeBlock,
+            controller: controller,
+            icon: Icons.code,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: showIndent,
-            child: IndentButton(
-              icon: Icons.format_indent_increase,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              isIncrease: true,
-            ),
+        if (isButtonGroupShown[2] &&
+            (isButtonGroupShown[3] || isButtonGroupShown[4]))
+          VerticalDivider(
+            indent: 12,
+            endIndent: 12,
+            color: Colors.grey.shade400,
           ),
-          Visibility(
-            visible: showIndent,
-            child: IndentButton(
-              icon: Icons.format_indent_decrease,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              isIncrease: false,
-            ),
+        if (showQuote)
+          ToggleStyleButton(
+            attribute: Attribute.blockQuote,
+            controller: controller,
+            icon: Icons.format_quote,
+            iconSize: toolbarIconSize,
           ),
-          Visibility(
-            visible: showQuote,
-            child: VerticalDivider(
-              indent: 12,
-              endIndent: 12,
-              color: Colors.grey.shade400,
-            ),
+        if (showIndent)
+          IndentButton(
+            icon: Icons.format_indent_increase,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            isIncrease: true,
           ),
-          Visibility(
-            visible: showLink,
-            child: LinkStyleButton(
-              controller: controller,
-              iconSize: toolbarIconSize,
-            ),
+        if (showIndent)
+          IndentButton(
+            icon: Icons.format_indent_decrease,
+            iconSize: toolbarIconSize,
+            controller: controller,
+            isIncrease: false,
           ),
-          Visibility(
-            visible: showHorizontalRule,
-            child: InsertEmbedButton(
-              controller: controller,
-              icon: Icons.horizontal_rule,
-              iconSize: toolbarIconSize,
-            ),
+        if (isButtonGroupShown[3] && isButtonGroupShown[4])
+          VerticalDivider(
+            indent: 12,
+            endIndent: 12,
+            color: Colors.grey.shade400,
           ),
-        ]);
+        if (showLink)
+          LinkStyleButton(
+            controller: controller,
+            iconSize: toolbarIconSize,
+          ),
+        if (showHorizontalRule)
+          InsertEmbedButton(
+            controller: controller,
+            icon: Icons.horizontal_rule,
+            iconSize: toolbarIconSize,
+          ),
+      ],
+    );
   }
 
   final List<Widget> children;
   final double toolBarHeight;
+  final bool? multiRowsDisplay;
 
-  @override
-  _QuillToolbarState createState() => _QuillToolbarState();
+  /// The color of the toolbar.
+  ///
+  /// Defaults to [ThemeData.canvasColor] of the current [Theme] if no color
+  /// is given.
+  final Color? color;
+
+  final FilePickImpl? filePickImpl;
 
   @override
   Size get preferredSize => Size.fromHeight(toolBarHeight);
-}
 
-class _QuillToolbarState extends State<QuillToolbar> {
   @override
   Widget build(BuildContext context) {
+    if (multiRowsDisplay ?? true) {
+      return Wrap(
+        alignment: WrapAlignment.center,
+        runSpacing: 4,
+        spacing: 4,
+        children: children,
+      );
+    }
     return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8),
-      constraints: BoxConstraints.tightFor(height: widget.preferredSize.height),
-      color: Theme.of(context).canvasColor,
-      child: CustomScrollView(
-        scrollDirection: Axis.horizontal,
-        slivers: [
-          SliverFillRemaining(
-            hasScrollBody: false,
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-              children: widget.children,
-            ),
-          ),
-        ],
-      ),
+      constraints: BoxConstraints.tightFor(height: preferredSize.height),
+      color: color ?? Theme.of(context).canvasColor,
+      child: ArrowIndicatedButtonList(buttons: children),
     );
   }
 }